Added server-side auth and input check.. - Cleaned up some logic for consistency/clarity; - Added/updated tests.

+++ backend/controllers/workoutController.js
@@ -2,6 +2,10 @@ const Workout = require('../models/workoutModel');
 const mongoose = require('mongoose');
 
 const getAllItems = async (req, res) => {
+  if(!req.user){
+    res.status(401).json({error: "Not authorized"})
+    return
+  }
  const page = req.query.p || 0;
  const itemsPerPage = 3;
  const search = req.query.search || null;
@@ -37,14 +41,22 @@ const getAllItems = async (req, res) => {
 };
 
 const addItem = async (req, res) => {
+    if(!req.user){
+      res.status(401).json({error: "Not authorized"})
+      return
+    }
     const {title, muscle_group, reps, load} = req.body;
+    if(!title || !muscle_group || !String(reps) || !String(load)){
+      res.status(400).json({error: "Please fill out the empty fields"})
+      return
+    }
     const user_id = req.user._id;
     const allWorkoutsByUser = await Workout.find({ user_id });
     const limit =
       process.env.NODE_ENV !== "test"
         ? Number(process.env.MAX_WORKOUTS_PER_USER)
         : Number(process.env.TEST_MAX_WORKOUTS_PER_USER);
-    if (allWorkoutsByUser.length >= limit) {
+    if (allWorkoutsByUser.length === limit) {
       const id = allWorkoutsByUser[0]._id;
       await Workout.findOneAndDelete({ _id: id });
     }
@@ -87,6 +99,10 @@ const deleteItem = async (req, res) => {
 };
 
 const deleteAllUserItems = async (req, res) => {
+  if(!req.user){
+    res.status(401).json({error: "Not authorized"})
+    return
+  }
   const user_id = req.user._id;
   try {
    const workouts = await Workout.deleteMany({ user_id });
